Do not crash on user related authentication errors
Currently if the authentication fails because of user related errors
(such as user cancelling the authentication process) an exception of
type `StatusAuthnFailed` is thrown during parsing the server response.
Since this exception isn't caught, it will crash the entire Django
application and display 500 Internal Server Error to the user. To avoid
this, catch the exception during parsing the response and deny access if
the exception is thrown.